Add new lookup controller
