+++ SignInCheckIn/SignInCheckIn/Controllers/EventController.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Web.Http;
 using System.Web.Http.Description;
-using MinistryPlatform.Translation.Repositories.Interfaces;
 using SignInCheckIn.Exceptions.Models;
 using SignInCheckIn.Models.DTO;
 using SignInCheckIn.Security;
@@ -263,42 +262,6 @@ namespace SignInCheckIn.Controllers
             catch (Exception e)
             {
                 var apiError = new ApiErrorDto($"Error updating Household ID of {householdDto.HouseholdId}", e);
-                throw new HttpResponseException(apiError.HttpResponseMessage);
-            }
-        }
-
-        [HttpGet]
-        [ResponseType(typeof(HouseholdDto))]
-        [VersionedRoute(template: "geStates", minimumVersion: "1.0.0")]
-        [Route("geStates")]
-        public IHttpActionResult GeStates()
-        {
-            try
-            {
-
-                return Authorized(token => Ok());
-            }
-            catch (Exception e)
-            {
-                var apiError = new ApiErrorDto($"Error getting states", e);
-                throw new HttpResponseException(apiError.HttpResponseMessage);
-            }
-        }
-
-        [HttpGet]
-        [ResponseType(typeof(HouseholdDto))]
-        [VersionedRoute(template: "getCountries", minimumVersion: "1.0.0")]
-        [Route("getCountries")]
-        public IHttpActionResult GetCountries()
-        {
-            try
-            {
-
-                return Authorized(token => Ok());
-            }
-            catch (Exception e)
-            {
-                var apiError = new ApiErrorDto($"Error getting countries", e);
                 throw new HttpResponseException(apiError.HttpResponseMessage);
             }
         }
