+++ SignInCheckIn/SignInCheckIn/SignInCheckIn.csproj
@@ -218,6 +218,7 @@
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Controllers\ChildCheckinController.cs" />
     <Compile Include="Controllers\ChildSigninController.cs" />
+    <Compile Include="Controllers\LookupController.cs" />
     <Compile Include="Controllers\EventController.cs" />
     <Compile Include="Controllers\HelloWorldController.cs" />
     <Compile Include="Controllers\KioskController.cs" />
